split components from workspace utils to separate ts files
